(doc) Rename components used for editing workout to first state field that it's editing, followed by "Edit" (ie. TimeEdit) to clarify that it is a component rather than a function.
