+++ client/scripts/common/editWorkoutComponents/holdEdit.js
@@ -16,8 +16,8 @@ var HOLD_CHOICES = [5,10,15,20,25,30,35,40,45,50,55,60,65,70,75,80,85,90,95,100,
 
 var PickerItemIOS = PickerIOS.Item;
 
-var EditHold = React.createClass({
-  /*EditHold state does not reflect store's state because each workout has many exercises which can be modified. If it did, the number of listeners would be too high. */
+var HoldEdit = React.createClass({
+  /*HoldEdit state does not reflect store's state because each workout has many exercises which can be modified. If it did, the number of listeners would be too high. */
   getInitialState: function() {
     return {
       showHoldSelection: false
@@ -68,4 +68,4 @@ var EditHold = React.createClass({
   }
 });
 
-module.exports = EditHold;
+module.exports = HoldEdit;
