+++ client/scripts/components/modifyWorkout/exerciseEdit.js
@@ -5,9 +5,9 @@ var modifyWorkoutStore = require('../../stores/modifyWorkoutStore');
 var modifyWorkoutActions = require('../../actions/modifyWorkoutActions');
 
 //Load components
-var EditReps = require('../../common/editWorkoutComponents/editReps');
-var EditLoad = require('../../common/editWorkoutComponents/editLoad');
-var EditHold = require('../../common/editWorkoutComponents/editHold');
+var RepEdit = require('../../common/editWorkoutComponents/repEdit');
+var LoadEdit = require('../../common/editWorkoutComponents/loadEdit');
+var HoldEdit = require('../../common/editWorkoutComponents/holdEdit');
 
 var {
   StyleSheet,
@@ -16,7 +16,7 @@ var {
 } = React;
 
 
-var EditExercise = React.createClass({
+var ExerciseEdit = React.createClass({
 
   render: function(){
     //Load props
@@ -32,10 +32,10 @@ var EditExercise = React.createClass({
     var renderRepsOrHold = function(exercise) {
       /* jshint ignore:start*/
       if(exercise.reps){
-        repsOrHold = <EditReps exercise={exercise} exerciseNum={exerciseNum} roundNum={roundNum}/>;
+        repsOrHold = <RepEdit exercise={exercise} exerciseNum={exerciseNum} roundNum={roundNum}/>;
       }else if(exercise.hold){
         //TO DO: enable user to edit hold and name on click
-        repsOrHold = <EditHold exercise={exercise} exerciseNum={exerciseNum} roundNum={roundNum}/>;
+        repsOrHold = <HoldEdit exercise={exercise} exerciseNum={exerciseNum} roundNum={roundNum}/>;
       }
       /* jshint ignore:end*/
     };
@@ -54,7 +54,7 @@ var EditExercise = React.createClass({
       //TO DO: enable user to edit load on click
       /* jshint ignore:start*/
       if(exercise.load.val) {
-        load = <EditLoad exercise={exercise} exerciseNum={exerciseNum} roundNum={roundNum}/>;
+        load = <LoadEdit exercise={exercise} exerciseNum={exerciseNum} roundNum={roundNum}/>;
       } else {
         load = null;
       }
@@ -77,4 +77,4 @@ var EditExercise = React.createClass({
   }
 });
 
-module.exports = EditExercise;
+module.exports = ExerciseEdit;
